Various minor improvements

+++ samples/core/copybuffer/main.cpp
@@ -100,7 +100,7 @@ int main(int argc, char** argv)
             if (!strcmp(argv[i], "-d"))
             {
                 ++i;
-                if (i < static_cast<size_t>(argc))
+                if (i < argc)
                 {
                     deviceIndex =
                         static_cast<cl_uint>(strtoul(argv[i], NULL, 10));
@@ -109,7 +109,7 @@ int main(int argc, char** argv)
             else if (!strcmp(argv[i], "-p"))
             {
                 ++i;
-                if (i < static_cast<size_t>(argc))
+                if (i < argc)
                 {
                     platformIndex =
                         static_cast<cl_uint>(strtoul(argv[i], NULL, 10));
